Update from PenRed/PenRed

+++ src/particleGen/specific/includes/PSFsource.hh
@@ -1,8 +1,8 @@
 
 //
 //
-//    Copyright (C) 2019 Universitat de València - UV
-//    Copyright (C) 2019 Universitat Politècnica de València - UPV
+//    Copyright (C) 2019-2020 Universitat de València - UV
+//    Copyright (C) 2019-2020 Universitat Politècnica de València - UPV
 //
 //    This file is part of PenRed: Parallel Engine for Radiation Energy Deposition.
 //
@@ -102,7 +102,7 @@ class psf_specificSampler : public abc_specificSampler<pen_particleState>{
   
   void sample(pen_particleState& state,
 	      pen_KPAR& genKpar,
-	      unsigned long& dhist,
+	      unsigned long long& dhist,
 	      pen_rand& /*random*/);
   
   int configure(double& Emax,
